requirements.txt
fastapi
uvicorn
redis
croniter
jinja2
python-multipart
python-jose[cryptography]
PyJWT